api/requirements.txt
flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
flask-cors==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0 